requirements.txt
matplotlib>=3.8.3
numpy>=1.24.4
opencv_contrib_python>=4.8.1.78
opencv_python>=4.8.0.76
opencv_python_headless>=4.10.0.84
Pillow>=10.4.0
skimage>=0.0